+++ src/Interfaces/Payment/Scheme/InstantPreferredSchemeSelectionInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TrueLayer\Interfaces\Scheme;
+namespace TrueLayer\Interfaces\Payment\Scheme;
 
 interface InstantPreferredSchemeSelectionInterface extends InstantSchemeSelectionInterface, SchemeSelectionInterface
 {