+++ src/Interfaces/Payment/Scheme/UserSelectedSchemeSelectionInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TrueLayer\Interfaces\Scheme;
+namespace TrueLayer\Interfaces\Payment\Scheme;
 
 interface UserSelectedSchemeSelectionInterface extends SchemeSelectionInterface
 {