+++ src/Interfaces/Payout/Scheme/PreselectedSchemeSelectionInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace TrueLayer\Interfaces\Scheme;
+namespace TrueLayer\Interfaces\Payout\Scheme;
 
 interface PreselectedSchemeSelectionInterface extends SchemeSelectionInterface
 {